(MAINT) Add go.mod for theme and correct site mod
Prior to this change, the go.mod for the site pointed to an incorrect URL
and the base theme did not have its own go.mod, which made it impossible
to use as a hugo module.

This change corrects both issues and ensures that fonts are treated as
binary files by git instead of as plaintext.