(MAINT) Reorganize repo  and site
This change heavily reorganizes the project files and the site layout.

The majority of the files for the website are moved into the `site`
folder to declutter the top-level as much as possible. The `content` and
`data` folders remain at the top-level for ease of access and because
they represent continual areas of contribution and work.

The site's layout has been completely reworked so that the nav menu
shows:

- The guides for getting started, using Platen, and theming Platen in
  the first block
- The modules in their own flattened section
- Tools, resources, and the contributing guide in the third block.

In this update, the module-specific documentation is also moved into the
docs folder of the module's source so that the docs live closer to their
implementation. This should help separate the reference/modules content
from the broader conceptual content.

With this scaffolding laid down, we should be in a better position to
begin filling these sections out with useful information.