(MAINT) Fix mobile view
Prior to this change, the mobile view for the `platen-header` had display set
to hidden instead of block in mobile.

This change adds the `!important` directive to the CSS for the header on mobile,
fixing the display.

It also adds `text-size-adjust` to the normalize CSS for compatibility, discovered
while working on fixing the mobile view.