(MAINT) Define Front Matter metadata for Toroidal
This change adds the `toroidal` and `toroidal/admin` content types to
the Front Matter configuration so that site maintainers can edit the
metadata for those pages in a friendly UI.

For member pages and the admin page, this is fully functional. For the
webring index, this is limited currently by the implementation of
estruyf/vscode-front-matter#362, which introduces `where` logic for the
fields but doesn't yet support using that logic to control behavior
based on the file name, which would allow us to handle section pages
(which are always named `_index.md`) separately from member pages.

This change also updates the archetypes for Toroidal to set the
`type` to `toroidal` for the index and member pages, ensuring that
Front Matter picks up the type correctly for selecting the metadata to
display. Due to a limitation outlined in estruyf/vscode-front-matter#360
(see [this comment for more information][01]), we have to use `type` to
identify files. This has some implications for Hugo sites which don't
_directly_ impact this change.

[01]: https://github.com/estruyf/vscode-front-matter/issues/360#issuecomment-1323708970