(MAINT) Enable features to extend site menu
Prior to this change, there were only three ways to add entries to the
site menu:

1. By having content in the `Platen.Display.RootMenuSection`, which
   automatically adds the pages and sections of that folder unless you
   override that behavior in a page or section's front matter.
2. With Hugo's `menu` configuration, you can add entries before and
   after the main section.
3. With the `platen/inject/menu-before` and `platen/inject/menu-after`
   partials, you can add any additional menu content you want.

This meant that the only way for modules extending Platen to add entries
was to ensure that the entries are sourced from content or to use the
injection partial. Using the injection partial prevents a site
maintainer or other module from using that injection, as the shadowing
logic overwrites partials.

This change introduces a new method for features to extend the menu,
building on the model introduced for titles and styles, with the
`Platen.Features.*.Partials` setting, adding handling for `Menu`.

In this implementation, a module providing a feature can specify a
partial to use in the menu rendering. For example:

```yaml
params:
  Platen:
    Features:
      MyFeature:
        Enabled: true
        Partials:
          Menu: myFeature/platen/foo
```

With that defined, site maintainers (and feature developers) can define
entries in the new `Platen.Display.Menu` key. It has six defined places
to add menu entries:

- `BeforeInjection` places the entries at the very top of the site menu,
  after the brand and search section.
- `BeforeConfig` places the entries after any from the injection partial
  but before the entries from the `menu.before` setting in Hugo's
  configuration file.
- `BeforeMain` places the entries after the entries from the
  `menu.before` setting in Hugo's configuration file but before the
  content added from the root content section by Platen.
- `AfterMain` places the entries after the content added from the root
  content section by Platen but before the entries from the `menu.after`
  setting in Hugo's configuration file.
- `AfterConfig` places the entries after the entries from the
  `menu.after` setting in Hugo's configuration file but before the
  injection partial.
- `AfterInjection` places the entries at the very bottom of the site
  menu after everything else.

In this implementation, each position can have its entries placed in a
single containing group or add each entry on its own. Containing groups
are defined with the `Group` key at the top level of the position field.
If the top-level `Group` key is omitted, each entry in `Entries` is
added in its own unordered list.

Entries may define the `Weight` field. All weighted entries are added
before any unweighted entries and are ordered from lowest to highest
weight. Entries may also define the `Group` key with their own nested
entries following the same logic.

Every non-group entry **must** define the `Feature` field, which informs
Platen as to which feature the entry relies on for building itself.
Only entries with usable features are added. Others are skipped with a
warning.

Usable features are those that:

1. Have `Enabled` set to `true`
2. Define a partial for `Partials.Menu`

Features can define any additional settings that make sense for them in
the `Menu` setting, including top-level defaults for all positions.

When the menu partial is called, it is passed a dictionary with:

- The current page context (accessible as `.Context`)
- The active position (accessible as `.Position`)
- The entry definition (accessible as `.Entry`)

For example, for the menu on the page `content/foo/bar.md` and from the
following configuration:

```yaml
params:
  Platen:
    Features:
      MyFeature:
        Enabled: true
        Partials:
          Menu: myFeature/platen/foo
    Display:
      Menu:
        BeforeMain:
          Apple: true
          Entries:
            - Feature: MyFeature
              Foo: first
```

The menu will call the `myFeature/platen/foo` partial after the entries
from the `menu.before` setting in Hugo's configuration file but before
the content added from the root content section by Platen with the page
context for `foo/bar.md`, the position `BeforeMain`, and the value of
the entry.

This change also introduces:

- A new utility partial, `getEnabledFeatures`, to make it easier to
  retrieve the enabled features as we continue to expand the integration
  points for Platen.
- The `platen/menu/feature` partial, which handles retrieving the
  position settings from the site configuration and processing it for
  the defined group and entries at that position.
- The `platen/menu/list-item` partial, which handles generating a list
  item for the Platen site menu, including flattening/collapsing groups
  and handling navigable and non-navigable items.
- The `platen/menu/features/group` partial, which handles the Group
  definition from `Platen.Display.Menu.<Position>` or a nested Group in
  an entry.
- The `platen/menu/features/entries` partial, which handles a list of
  entries defined in `Platen.Display.Menu.<Position>` or in a nested
  Group entry.
- Updated schematization for these settings

Finally, this change reworks the folder structure for menu-related
partials, adding the `layouts/partials/platen/menu` folder and moving
the partials, except for the base `menu` partial, into that folder.