(MAINT) Reorganize footer & fix "Edit this page"
This change refactors the footer template, breaking its component
parts into standardized partials:

- `platen/footer/lastEditedOn` - Render the "Last edited on" link
- `platen/footer/editThisPage` - Render the "Edit this page" link
- `platen/footer/codeClipboard` - Add the click-to-copy functionality
  for code blocks.

Additionally, this change reworks the logic for the "Edit this page"
link to support mounted content and content directories with relative
paths, which previously broke the link building.

It introduces new settings in `Platen.Repository` and schematizes them:

- `ContentRoot` - An explicit override for the folder segment to use for
  the content directory in building the link. If this value isn't set,
  the partial uses the value of Hugo's `contentDir` parameter and trims
  leading periods and forward slashes.
- `MountedPaths` - Adds support for building URLs for mounted content,
  which the previous implementation couldn't manage. Supports ignoring
  the link for mounted content, linking to external repositories, and
  mapping to the correct folder in the current repository.

Finally, this change updates the site configuration to take advantage
of the new configuration options to fix the "Edit this page" links on
the site.