(MAINT) Reorganize and enhance header
This change reorganizes the header partial, splitting it into several
components:

- `platen/heaader/menuControl` - Renders the label for the input element
  to display the site menu in mobile view.
- `platen/header/title` - Renders a title element for the page as either
  `<strong>` or an `<h1>`.
- `platen/header/tocControl` - Renders the label for the input element
  to display the table of contents in mobile view.

This change adds new functionality to the module for rendering a page's
title as an `<h1>` instead of always in a `<strong>` element with the
new `Platen.Display.Header.TitleAsHeading` setting in the site config
and the `Platen.TitleAsHeading` setting in content frontmatter.

It updates the styling for the site to always render the `<h1>` heading,
even on mobile, and ensures the style continues to work for the old
rendering style.