(MAINT) Refactor TOC & fix related schemas/config
This change refactors the TOC to standardize it with the other rendering
partials and moves the generate partial from the utils folder to
`platen/toc/generate`, which is a more semantically correct location.

It also updates the schemas for the existing settings and the Front
Matter configuration definitions to match the implementation.