(MAINT) Add utils to prettify rendered HTML
This change adds the `getPrettyRender` utility partial, which
renders Markdown reliably and supports rendering as a block or
inline as well as handling leading whitespace. This makes the
HTML for rendered examples and generated source easier to read.

It also updates the `getSafeAttributes` utility partial to break
across multiple lines if the line length is too long or there are
too many attributes to read.