(MAINT) Rework KaTeX support
This change updates KaTeX to the latest version (0.16.4) and adds
a render hook for it as a companion to the shortcode.

This update also adds support for chemistry equations.

It refactors the implementation to pull KaTeX from the CDN at
build time, defaulting to 0.16.4 and updating the fonts to match.

This required updating the site configuration to include Katex as
a feature. By default, it is enabled with its version set to 0.16.4
and set to only load when used.

This update also adds the Features setting to page frontmatter with
the Katex setting group. Currently the only option is to enable KaTeX
on the page, which enables an author to use inline the KaTeX  wrapping
tags without the shortcode or codeblock.

In the future, these settings may support site-wide and page-only macros
and other options.