(MAINT) Rework mermaid support
This change removes the mermaid shortcode and replaces it with
a render hook, allowing an author to write a codeblock with the
`mermaid` language ID and have it transformed into a diagram.

It updates the default version to 9.3.0 and supports adding the
diagram in a figure or standalone as well as the new initialize
directives.

It reworks the implementation to pull Mermaid from the CDN at
build time, no longer vendoring the code.

This also required updating the site configuration to add Mermaid
as a feature. By default, it is set to enabled with the version
set to 9.3.0.

In the future, we may add support to the site configuration and
page frontmatter for defining initialization configuration merging.

This implementation adds commented-out support for mindmaps, which
are not working for static sites at this time, but can be added
back in for experimentation.

The configuration JSON is moved over to the assets folder, which
lays the groundwork for converting it into a partial in the future.