(MAINT) Update search implementation
This change updates the search implementation, moving the flexsearch
library into the assets folder. In the future, this should be updated
to pull flexsearch from the CDN. However, that will require major
refactoring to support the new major version.

This change brings it more inline with the rest of the dependency handling
for the module.