(MAINT) Reimplement columns shortcode
This commit reimplements the `columns` shortcode to match the implementation of
the `tabs` shortcode, using an inner `column` shortcode instead of a string split.

Now, instead of defining a ratio on the opening shortcode, authors can define the
`grow` parameter for the nested `column` shortcode to control its behavior.

Both the containing column group and each column can have classes added to them.

The implementation is also updated to support trimming leading whitespace, making
it easier to discern the nested content without numerous extra newlines.