(MAINT) Prettify and humanize button shortcode
This commit updates the implementation of the button shortcode
to prettify the rendered HTML and to humanize the parameters,
replacing `rel` with `relative`.

In the future, we may want to extend the behavior to support
a render hook for this functionality so authors can define a
normal link that acts as a button.

We should also add support for actual buttons that can take
actions.