(MAINT) Update itchio shortcode
This commit reworks the implementation of the itchio shortcode
to support both simple and complex use cases as well as adding
the title attribute for accessibility.

The simple usage is similar to the previous implementation, but
without control over styling. Complex usage requires named params
but now supports the full range of styling offers the itch API
supports.

In the future, we should support defining itch embeds in data and
merely referencing them in Markdown to make reuse easier.