(MAINT) Style and prettify art shortcode
This commit updates the implementation of the art shortcode
to prettify the rendered HTML and properly implement the
content warning behavior.

Now, when a content warning is specified, the image is blurred
by default and displays a content warning. Clicking on the image
unblurs it and hides the content warning.

In the future, we should support site-wide content warnings and
enable site visitors to configure their preferences so they
can unhide content by default.

This change also updates the site configuration to use the style
module for the shortcode.