(MAINT) Remove properties heading from schematize
This change updates the `schematize` shortcode to remove the
"Properties" heading from the rendered content. This heading
necessarily increased the heading level by 1 for all object
properties and made it extremely difficult to nest objects
in the same page, leading to an unnatural breakdown of the
definitions.

An object (H1) had properties defined in H3s beneath the
"Properties" H2. If that property was also an object, the
shortcode could render those properties under an H5 after
its own "Properties" H4. However, no further nesting for
objects was possible as the properties would then need to
be in an H7, effectively limiting us to an object with one
level of nesting.

With this change, schematize can render an object (H1) with
properties defined in H2s. If one of those  properties is an
object, its properties are then H3s. This expands the maximum
nesting level for objects from one level to four.