(MAINT) Prettify and clean up schematize docs rendering
This commit standardizes the rendering for the docs strings in schematize to
reflect the lessons learned in module development, making the rendering
easier to follow and more deterministic, no longer relying on the outputting
raw strings in the template for spacing.

It also prettifies the code to make it easier to read and follow, and removes
unused variable declarations.