(MAINT) Implement disabled shortcode erroring
This change upholds the note in the documentation that says
using disabled shortcodes causes Platen to raise an error.

It adds the new `getRegisteredShortcodes` partial in the
`configuration` folder to find and return all registered
shortcodes from the configuration. It also supports
retrieving only disabled shortcodes in this iteration.

The new `checkShortcodes` partial is added to the same
folder and called from the `htmlHead` partial oncer per
page. It retrieves the registered shortcodes that are
disabled, then loops over them to build a regex pattern
for the shortcode and its aliases. It looks through the
raw Markdown for the page to see if the shortcode is used
and, if it is, raises a descriptive error.

In the future, it may be worthwhile to allow users to
enable and disable shortcodes on a per-page basis.

Note that while this partial supports aliases for registered
shortcodes, we do not yet have a pattern for defining them
properly without copying the shortcode implementation.