(MAINT) Support empty attributes for renderers
Prior to this commit, the `memo/renderer/attribute` shortcode threw an
error if the memo data didn't define any attributes when merging the
common attributes with the defined ones.

This commit updates the logic to default the attributes to an empty map
if they're not specified. This enables the use of render hooks that only
use the common attributes.