(GH-18) Convert `columns` shortcode to codeblock
Prior to this change, authors could use the `columns` shortcode to add
content to a page separated into columns with flexible widths. This
change replaces the shortcode with a codeblock render hook.

It updates the documentation and configuration schema to reflect the
change and solidifies the use of a YAML block of options for Platen
codeblock render hooks.