(MAINT) Standardize katex codeblock renderer
This change updates the KaTeX codeblock render hook to use the standard
attribute-and-yaml-options implementation. It also updates the
implementation to render the markup in a div instead of a span and to
support adding the formula in a figure, like the mermaid codeblock.

It makes minor fixes to related partials to render the docs correctly.