(MAINT) Fix id-to-name munge for tabs
This change fixes the munging behavior for converting a tab id into
the tab name for the `tab` codeblock render hook. This only affects
tabs where the ID is specified without the name. It ensures that the
name replaces dashes and underscores with spaces before titleizing
the ID.

The ensures `some-id-value` becomes `Some ID Value` instead of the
incorrect `Some-ID-Value`.