(GH-24) Replace `button` shortcode with render hook
This change replaces the `button` shortcode with an image link render hook
for a more normal Markdown authoring experience. The image link render
hook was chosen over a normal link render hook because:

1. It adds an inline block element.
1. Link render hooks can not have attributes, but images can. This allows
   a user to set custom classes on their buttons.

This change also paves the way for adding presets to buttons and for
defining a more complex button implementation that can call javascript.