(MAINT) Reorganize schema/docs for Toroidal
This change reorganizes the schema data and documentation for
the Toroidal module by moving all of the docs into the `_docs`
folder. It also moves the docs for the config schemas from
`content/schemas` to `_docs/config` so the configuration info
will live with the rest of the module. Finally, it moves all
of the configuration schemas from the site-level data folder
into `_schema_data/schemas/Toroidal`.

This move is for two uses:

1. Using the `_` prefix for the folder names makes it clearer
   that they are not part of the standard layout for the module.
1. Moving the docs and schemas fully into the module ensures that
   all of what you need to look at when working on it is in one
   folder.

This is part of an overhaul for how the modules and config are
structured in preparation for the next release.