(MAINT) Reorganize platen docs/schemas
This change moves the documentation and schema files for Platen
into the `_docs` and `_schema_data` folders respectively. This
completes the migration for the extant modules.

Now we can begin properly restructuring them so that we can make
the documentation and usability better.

Ideally, we will not need to double-publish the schemas in the
module folder and again in the schema root folder.

We should also take the time to rename the properties, downcasing
and snake_casing them so we do not need to worry about casing
in user configs and partials.

We should also consider consolidation and other improvements where
possible.