(GH-40) Support merging site params from data
Prior to this change, the only way to apply configuration options for
the modules was to use Hugo's configuration files or environment
variables to define the settings.

The primary downside to this for site maintainers is that the `params`
key in the configuration can't be split across multiple files. Given the
number of configuration options Platen has, that could make configuring
a site overwhelming, even with the schematized configuration options
enabling intellisense and validation.

This change enables site maintainers to define their options in the
`_params` folder inside their data directory. With this, maintainers can
now split their Platen (and Schematize/Toroidal) configuration settings
across multiple files for less cognitive overhead.

Combined with Front Matter, in the future we'll be able to provide an
excellent UX for site maintainers where they can use a GUI to edit their
site configuration with autocomplete and validation. For now, they have
a similar experience with the files using the JSON schemas in VS Code.

This implementation provides three new partials usable by compatible
modules:

- `platen/param/getMergedValue` retrieves the deeply merged site
  configuration params, merging the values defined in `data/_params`
  over those in `site.Params`.
- `platen/param/getKey` retrieves a specific key from the deeply merged
  configuration params.
- `platen/param/resolve` resolves specific keys from the deeply merged
  configuration params and the page's front matter, optionally returning
  a default value if neither the site nor page define the key.

This change updates Platen to always use this functionality and the
Toroidal and Schematize modules to use it if Platen is available.

It also pre-defines the data files in the modules for easy copying and
for this site to use the local server for schemas to enable testing and
development.

Finally, this change takes the opportunity to update several partial
calls to cache them to reduce build time.

- Resolves #40