(MAINT) Update configurations and use data config
This change takes advantage of the merge-from-data functionality to
migrate the site's parameters from the configuration file to data files.

It also updates and sets the minimum hugo version for the modules and
migrates the hugo configuration for markup from the site to Platen and
uses the `_merge` key in the site to merge it in.

Finally, it deletes the `theme.(toml|yaml)` files as they're unused.