(MAINT) Publish initial FrontMatter configs
This change takes advantage of the improvements to the memo module to
publish an initial group of configs for FrontMatter users.

This initial pass doesn't do any further work with the documentation, it
just makes the composed configurations available for testing.

This is in support of estruyf/vscode-front-matter#407, which adds
the `frontMatter.extends` key.