(GH-25) Add initial support for data configs to memo
This change adds basic support for having memo publish Front Matter
configs that reference data types defined as JSON schemas, re-using
the path resolution from the configs. This initial implementation
replaces the `schema` key in the `frontMatter.data.*` settings if
that value is a scalar instead of a map or array.

This implementation is brittle but functional for current usages. It
does not (yet) resolve any `$ref` keys, which will not work in FM.
A future implementation will need to resolve those keys to their
values in the same way that the current implementation munges the
`schema` key.

This enables us to support presets and begin to move towards adding
the configuration and data settings to Front Matter.