(GH-25) Add presets for itch embeds
This change adds support for defining presets for itch embeds to
the markup by specifying a string instead of digit value for the
ID. It also schematizes the preset and includes it in the front
matter configuration builder.

Finally, it makes the full set of useful Front Matter settings for
the markup available at `/frontmatter/platen/markup/itch.json`.