(MAINT) Rework memo examples to be figures
This change converts the rendering for examples from the memo module
to write the code representations in figures with captions instead of
as sections under a heading.

This makes for a shorter TOC while keeping the examples linkable, as
each figure is given a unique ID matching the previous heading.