(GH-41) Implement body class injection
Prior to this change, Platen had unused code for injecting CSS classes
into the body of pages based on page parameters and parsing ancestor
pages to assemble a list of classes to inject.

This change reimplements the functionality in a more configurable way:

- The new `extra_classes` parameter in content front matter adds any
  listed classes to the `body` element of the page.
- The new `display.section_classes` setting in the site configuration
  is checked when rendering a page to see if the page's URL matches any
  defined setting, adding their classes to the injection list.

This makes things a little less magic and more understandable while
keeping the intended functionality.

- Resolves #41