(MAINT) Add markup for code snippets and key combos
This change stealth-adds two new pieces of markup that will
need to be documented eventually: `code_snippets` for inline
code highlighting and reusable code blocks, and `key_combos`
for inline and defined hotkeys without HTML.

These are required for documentation and serve the broader
user base, but are low-priority for user-facing documentation.

A future commit should add full schema and documentation for
them.