(FEAT) Add `sup` and `sub` markup
Prior to this change, there was no way to add text as superscript or
subscript except to use KaTeX markup. That's more correctly used for
mathematical or scientific markup, whereas superscript and subscript
have other purposes.

This change adds two new markup options to Platen for rendering text as
superscript or subscript.