(MAINT) Add styles/presets to `sub` markup
Prior to this change, the subscript markup didn't support pre-defined
styles or presets. With this change, the markup has been modified to
support both.

You can see how it works with the following example:

```markdown
This is some![sub:Subscript **text**]().

This uses an![sub:](preset:example "ope").

This is a![sub:styled option](baz).

This just adds a![sub:class list](. "foo bar")
```