(MAINT) Add styles/presets to `sup` markup
Prior to this change, the superscript markup didn't support pre-defined
styles or presets. With this change, the markup has been modified to
support both.

You can see how it works with the following example:

```markdown
This is some![sup:Subscript **text**]().

This uses an![sup:](preset:example "ope").

This is a![sup:styled option](baz).

This just adds a![sup:class list](. "foo bar")
```