(MAINT) Standardize and document styled_text
This change uses `styled_text` for the markup for semantic clarity. This
change uses the same structure as the markup for subscripts and
superscripts but for spans.

In the future, it might be worth extending this with a codeblock markup for
divs of text instead of spans.

This change includes schematization and documentation.