(MAINT) Support details markup as collapsible sections
This change enables using the `details` markup to create collapsible sections
by setting the `heading_level` for the markup to 1--6. By default, the summary
text for `details` is added as rendered Markdown inside the summary tag. With
this change, that rendered text is wrapped in a heading so it can show up in
the table of contents.

This change also makes some minor improvements to the styling for the `details`
elements:

- Replaces the default caret with a `+`/`-` toggle that floats to the end of
  the text, respecting language direction.
- Ensures all elements rendered inside the summary are rendered as inline blocks
  and without margins, ensuring nested heading elements do not break the flow.