(MAINT) Enable specifying font import URLs in config
Prior to this change, adding any fonts required a user to:

1. Download the fonts and add them to their `static` folder.
2. Create a new SCSS module for their fonts, like `assets/styles/fonts/_foo.scss` that
   has the `@font-face` definitions.
3. Update the `platen.theme.fonts.styles` setting to include the import name of the new
   SCSS module for their font.

This change adds a shortcut handler to the existing `font.styles` setting for fonts that
can be imported from the web. When the value for the `import_name` or simple string is
a URL (that is, it starts with `http://` or `https://`), Platen automatically constructs
the SCSS import statement for it.

This enables users to simplify their process from the above list to:

1. Update the `platen.theme.fonts.styles` setting to include the import URL for their font.