(MAINT) Fix shoelace integration with theme mode
Prior to this change, the shoelace feature's integration with Platen was
brittle and did not respect the theme's mode configuration.

This change pulls the shoelace CSS into the theme SCSS, adding on a
per-mode basis, and ensures that it can react to changes in the user's
preference.

These changes were necessary for beginning the work of more fully
incorporating shoelace into the module's theming.