(MAINT) Update Mermaid to v10 and require v10+
This change overhauls the integration with mermaid, ensuring that the
configuration is defined so elements aren't immediately rendered.

Instead, the init script registers observers for the elements with the
`mermaid` class and calls `mermaid.run` against them only when they are
first actually visible. This is to support using web components like
those added by shoelace, which may not be correctly handled by mermaid
if rendered while they're not visible.

Doing so required writing a handler, which has been implemented based
on the Mermaid v10 API where `mermaid.run` is an async function. Instead
of contorting Platen to support older versions of mermaid, we decided
to standardize on this new API, even though it's a breaking change from
older versions. We expect users mostly aren't overriding the Mermaid
version at this point, so they shouldn't notice.

In implementing this change, it also became necessary to provide a
Platen javascript module for the site. Any other script module that
needs to use its functions, like `platen.onFirstVisible()`, can get
the url to the script for importing by calling the new partial
`platen/utils/getModule`:

```javascript
{{- $PlatenModule := partialCached "platen/utils/getModule" $Context -}}
import * as platen  from '{{ $PlatenModule.RelPermalink }}'
```