(MAINT) Fix `getMergedValue` for falsy values
Prior to this change, the `platen/param/getMergedValue` partial couldn't
correctly apply falsy values when merging maps. The logic used the
`with` conditional, which only runs when the value is truthy.

This change corrects the implementation to use the `isset` function
instead of `with`, ensuring that _any_ set value in the merging map
is applied over the existing value in the base map.

This ensures that users can actually set values which have `true` as
their default to `false`.