(MAINT) Fix logo display
Prior to this change, a bug in how we display our logo
was causing the colors to re-invert on pages where the
shoelace theme was loaded but in light mode. This change
ensures we load shoelace on every page, which we will
do eventually when the main components swap to using
shoelace.