(SQUASH) Simplify details codeblock markup
This change reuses the established merging pattern from image link markup
render hooks for codeblocks. It defined the new `getMergedCodeblockOptions`
utility partial for merging the YAML options and codeblock attributes and
moves the logic for canonicalizing the codeblock options into a private
utility function for the markup.

The goal is to make it simpler to implement the markup render hooks by
standardizing their layout and using shared utility partials wherever
sensible.

In this model, we use utility partials for:

- Getting the list of valid language IDs
- Retrieving the markdown/yaml options from the codeblock contents
- Merging the values from the markup attributes with yaml options

Every markup will need to implement:

- Its own `_canonicalize` utility partial, for merging in the
  settings from configuration, defined defaults that are not from
  the configuration, performing specific value munging, selecting
  the template to use for rendering, and providing any values those
  templates need for rendering.
- At least one rendering template that receives the canonicalized
  information as its parameters.

Markup implementations may:

- Support multiple templates
- Support users defining their own custom template