(MAINT) Rework Markup
This is a major refactor, completely redesigning the markup
implementation around lessons learned from the past several months.

It brings the full set of (publicly published) markup into the
standardized `markup -> _canonicalize -> template` flow and improves or
adds utility partials to reduce code where possible.

It adds the new `_scaffold` folder into the layouts folder to serve as
a base template for new markup and hopefully ease the contributions a
bit.

Future changes will need to be made to continue to improve the module
and make is easier to contribute markup, but this improves things
enormously.