(MAINT) Add `standardize_height` to tabs
Prior to this change, there was no convenient way to make sure that tabs in a group
had their height equalized. When the height for tabs in a group varies, switching
between the tabs causes the page layout to shift, which can be confusing or
frustrating for readers.

This change adds the `standardize_height` option to tab groups and the site
configuration, enabling users to ensure tab heights are standard in every or any tab
group on their site. It updates the documentation and schema.

This change also renames the script for tabs from `handleClosing.js` to `handler.js`
to better reflect the script purpose now that it does more than just handle `sl-close`
events for tab groups.

The implementation waits until the tab groups are all loaded on a page, then loops over
the tab groups defined with the `data-standardize-height` attribute set to `true`. For
each tab group, it loops over the panels to find the `offsetHeight` of the tallest panel,
then sets that as the height for every panel.