(MAINT) Add initial icon support
This change adds initial support for icon markup and registering
alternate icon libraries to Platen. This feature is not finished,
as the icon options for other markup are not aware of it and the
documentation is not written.

It simplifies the usage of icons across a site and enables users
to load from a long list of well-known icon libraries.

In the future, we could offer icon vendoring at build-time and
potentially generate validation info, so users can know whether
the icon name they specify exists for that library.

For now, it simply trusts the author to use the correct name.

It also has custom handling for specifying variants and libraries
in the icon name, using the format:

```
`<name>&<variant>@<library>
```

Where `variant` and `library` are optional. Users can set an
alternate default library and specify the title input on their
image link to add an accessibility label for the icon.

More work is needed, especially around documentation and thinking
about custom icon libraries.