(MAINT) Fix schematize ref resolution
Prior to this change, the logic for the schematize resolver for data
only worked when the base URL pointed to the base of a website, like
`https://foo.com`, not `https://foo.com/bar`.

When the base URL had a suffix after the domain name, the partial
resolved the relative path to the file as `/suffix/...` instead
of `/...`, which the `site.GetPage` function could not resolve,
since the content does not live in a `suffix` folder in the
content.

This change munges the prefix for site-relative paths in the
resolution to use `/` as the prefix, even when the site is
hosted in a folder on the domain.