(MAINT) Update param resolution for multi-key lookup
This change reworks the implementation of the `platen/param/*` partials to handle
receiving multiple keys to resolve in the site configuration and page front matter.

Prior to this change, resolution was single-key only, which meant that handling
renamed keys or multi-layer defaults was difficult and unintuitive.

This change allows a smoother transition model for deprecating and renaming keys.

This change also:

- Updates the documentation for the Param partials to clarify their use and behavior.
- Fixes a bug in the code_snippet image render hook handler, which was always using
  an invalid lookup for the markup configuration.
- Fixes a would-be bug in the resolution for toroidal classes, which might not be set
  and where resolution did not specify a default value.