(MAINT) Replace footer links with buttons
This change reimplements the editThisPage and lastEditedOn partials to render
`<sl-button>` elements with the `text` variant instead of button-styled links.

Both are configurable, using the `edit_this_page` and `last_edited_on` settings
under `platen.display.footer`.

The last edited on button shows the time elapsed since the last commit by
default and has a tooltip with more information. Clicking the button opens that
commit in the source repository.

The edit this page button links to the edit UI for that page in the default
branch of the project's source repository.

Both buttons can be disabled. Users can set `use_legacy` for the button's to
`false` to continue to use the old links for now.

The button icons and variants are configurable as well.

This change doesn't include the docs, which will be added in a schema rework
effort in the near future.